When opening a repository let libgit2 find the appropriate git directory
This is required when the repository's submodules
have been initialised with git 1.8, and their .git
is nested within the root project's.

If opened directly through the git directory,
libgit2 will lose track of which working directory
a user is trying to manipulate; will be unable to
find refs, and other such fun.